Refactored infrastructure views
