+++ infrastructure/views/__init__.py
@@ -1,0 +1,2 @@
+from .api import *
+from .templates import *
